moved sampler scripts
